Rename: Change folder name to modules -> molecules, organisms
