+++ src/components/templates/GnbTemplate.tsx
@@ -1,4 +1,4 @@
-import { Gnb } from '@/components/modules/Gnb'
+import { Gnb } from '@/components/organisms/Gnb'
 
 export interface GnbTemplateProps {
   children: React.ReactNode
@@ -9,7 +9,7 @@ const GnbTemplate = ({ children }: GnbTemplateProps) => {
     <div>
       <Gnb />
       <div className="size-full transition-[padding-left] duration-500">
-        <main className="px-6 py-16">{children}</main>
+        <main className="mx-auto max-w-7xl px-6 py-[4.5rem]">{children}</main>
       </div>
     </div>
   )
